+++ mujoco_extensions/mujoco_3d_lidar/docs/lidar_360x1.jpg
