+++ mujoco_extensions/mujoco_3d_lidar/docs/lidar_45x1.jpg
