+++ mujoco_extensions/mujoco_3d_lidar/example/lidar.xml
