+++ mujoco_extensions/mujoco_3d_lidar/example/scene.xml
