+++ mujoco_extensions/mujoco_3d_lidar/src/3dlidar.cpp
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-#include "mujoco_lidar_plugin/3dlidar.h"
+#include "mujoco_3d_lidar/3dlidar.h"
 
 #include <algorithm>
 #include <cctype>
